requirements.txt
streamlit
streamlit-drawable-canvas
google-generativeai
pillow
numpy
pyttsx3
